fix: refactor and reinstated sidebar custom instructions panel
